Create main.py
Uses the pydexcom library and the standard sqlite library to download CGM data into a database for personal backup. 

get_glucose_readings function is given data for a time period a little over present to 24 hours prior to request. This results in a little over one day of data. 

This script will download the data and delete the duplicates. Probably best used as a daily scheduled cron job.